[plugin-admin] store is now sliced in global and pluginstore
