+++ ui/plugin-admin/src/components/StoreProvider.js
@@ -22,12 +22,17 @@ export const useApiEndpoint = () =>
   useAppStore((state) => state.globals.apiEndpoint)
 export const useLoggedIn = () => useAppStore((state) => state.globals.loggedIn)
 export const useAuthData = () => useAppStore((state) => state.globals.authData)
-export const usePluginConfig = () =>
-  useAppStore((state) => state.globals.pluginConfig)
-export const useShowDetailsFor = () =>
-  useAppStore((state) => state.globals.showDetailsFor)
 
 export const useGlobalsActions = () =>
   useAppStore((state) => state.globals.actions)
+
+/// Plugin
+export const usePluginConfig = () =>
+  useAppStore((state) => state.plugin.pluginConfig)
+export const useShowDetailsFor = () =>
+  useAppStore((state) => state.plugin.showDetailsFor)
+
+export const usePluginActions = () =>
+  useAppStore((state) => state.plugin.actions)
 
 export default StoreProvider
