+++ ui/plugin-admin/src/lib/store/index.js
@@ -6,10 +6,12 @@
 import { createStore } from "zustand"
 import { devtools } from "zustand/middleware"
 import createGlobalsSlice from "./createGlobalsSlice"
+import createPluginSlice from "./createPluginSlice"
 
 export default () =>
   createStore(
     devtools((set, get) => ({
       ...createGlobalsSlice(set, get),
+      ...createPluginSlice(set, get),
     }))
   )
